Remove the dependency on the framework directory
This was previously a revert of commits
f9bbe0de4c97ca532d052106bb471756e55742ae
and
469f7811fac98617e544aac2b1b2db56040c95db
but due to conflicts that came up because of
later commits in Mbed TLS, turning that into
a single commit to revert the changes manually.

The framework directory was introduced in Mbed TLS 3.6.0
and for now doesn't do much more than make CMake fail if
`git submodule update --init` is not run.

This needs to be revisited once bigger changes
are made (with Mbed TLS 4.0) and the whole
repository structure changes.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>